Quick_Meet-main/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-apps-meet==0.1.6
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
openai==0.28.0
outcome==1.3.0.post0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
PySocks==1.7.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.1
selenium==4.19.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.29
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
yarl==1.9.4
